+++ lib/flutter_thermal_printer_platform_interface.dart
@@ -29,7 +29,7 @@ abstract class FlutterThermalPrinterPlatform extends PlatformInterface {
     throw UnimplementedError("connect() has not been implemented.");
   }
 
-  Future<void> printText(Printer device, Uint8List data) {
+  Future<void> printText(Printer device, Uint8List data, {String? path}) {
     throw UnimplementedError("printText() has not been implemented.");
   }