+++ macos/Classes/FlutterThermalPrinterPlugin.swift
@@ -4,33 +4,68 @@ import IOUSBHost
 import IOKit
 import IOKit.usb
 import Foundation
-import USBDeviceSwift
- 
-public class FlutterThermalPrinterPlugin: NSObject, FlutterPlugin , USBWatcherDelegate, FlutterEventChannel{
-    public func deviceAdded(_ device: io_object_t) {
-            print("device added: \(device.name() ?? "<unknown>")")
-            if let usbDevice = device.getInfo() {
-                print("usbDevice.getInfo(): \(usbDevice)")
-            }else{
-                print("usbDevice: no extra info")
-            }
-        }
+import ORSSerial
+
+public class FlutterThermalPrinterPlugin: NSObject, FlutterPlugin  , FlutterStreamHandler{
     
-    public func deviceRemoved(_ device: io_object_t) {
-            print("device removed: \(device.name() ?? "<unknown>")")
-        }
-    private var usbWatcher: USBWatcher!
-    
-    override init() {
-        super.init()
-        usbWatcher = USBWatcher(delegate: self)
+    private var eventSink: FlutterEventSink?
+    public func onListen(withArguments arguments: Any?, eventSink events: @escaping FlutterEventSink) -> FlutterError? {
+        eventSink = events;
+        return nil
     }
- 
+    
+    public func onCancel(withArguments arguments: Any?) -> FlutterError? {
+        eventSink = nil
+        return nil
+    }
+    
+//     public func deviceAdded(_ device: io_object_t) {
+//             print("device added: \(device.name() ?? "<unknown>")")
+//             if let usbDevice = device.getInfo() {
+//                 print("usbDevice.getInfo(): \(usbDevice)")
+//                 // Map usbDevice to a string and send it to the Flutter side
+//                 let deviceDict = [
+//                     "type" : "Added",
+//                     "device": usbDevice.toDictionary()
+//                 ] as [String : Any]
+//                 if eventSink != nil {
+//                     eventSink!(deviceDict)
+//                 }
+//             }else{
+//                 print("usbDevice: no extra info")
+//             }
+//         }
+//    
+//     public func deviceRemoved(_ device: io_object_t) {
+//             print("device removed: \(device.name() ?? "<unknown>")")
+//             if let usbDevice = device.getInfo() {
+//                 print("usbDevice.getInfo(): \(usbDevice)")
+//                 // Map usbDevice to a dictionary and send it to the Flutter side
+//                 let deviceDict = [
+//                     "type" : "Removed",
+//                     "device": usbDevice.toDictionary()
+//                 ] as [String : Any]
+//                 if eventSink != nil {
+//                     eventSink!(deviceDict)
+//                 }
+//             }else{
+//                 print("usbDevice: no extra info")
+//             }
+//         }
+//     private var usbWatcher: USBWatcher!
+//    
+//     override init() {
+//         super.init()
+//         usbWatcher = USBWatcher(delegate: self)
+//     }
+// 
     
   public static func register(with registrar: FlutterPluginRegistrar) {
     let channel = FlutterMethodChannel(name: "flutter_thermal_printer", binaryMessenger: registrar.messenger)
     let instance = FlutterThermalPrinterPlugin()
     registrar.addMethodCallDelegate(instance, channel: channel)
+    let eventChannel =  FlutterEventChannel(name: "flutter_thermal_printer/events", binaryMessenger: registrar.messenger)
+    eventChannel.setStreamHandler(instance)
   }
 
   public func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
@@ -38,7 +73,23 @@ public class FlutterThermalPrinterPlugin: NSObject, FlutterPlugin , USBWatcherDe
     case "getPlatformVersion":
       result("macOS " + ProcessInfo.processInfo.operatingSystemVersionString)
     case "getUsbDevicesList":
-        result(getAllUsbDevice())
+        var devices = Array<String>()
+        let matchingDict = IOServiceMatching(kIOUSBDeviceClassName)
+        let iter = UnsafeMutablePointer<io_iterator_t>.allocate(capacity: 1)
+        let kr = IOServiceGetMatchingServices(kIOMasterPortDefault, matchingDict, iter)
+        if kr != KERN_SUCCESS {
+            print("Error getting services")
+            result(devices)
+        }
+        while case let usbDevice = IOIteratorNext(iter.pointee), usbDevice != 0 {
+            let bsdPath = getBSDPath(for: usbDevice)
+            print(bsdPath as Any)
+            let device = usbDevice.getInfo()
+            var deviceInfo = device?.toDictionary()
+            deviceInfo?["bsdPath"] = bsdPath
+            devices.append(String(data: try! JSONSerialization.data(withJSONObject: deviceInfo as Any, options: .prettyPrinted), encoding: .utf8)!)
+        }
+        result(devices)
     case "connect":
         let args = call.arguments as? [String: Any]
         let vendorID = args?["vendorId"] as? String
@@ -49,173 +100,63 @@ public class FlutterThermalPrinterPlugin: NSObject, FlutterPlugin , USBWatcherDe
         let vendorID = args?["vendorId"] as? String ?? "0"
         let productID = args?["productId"] as? String ?? "0"
         let data = args?["data"] as? Array<Int> ??  []
-        printData(vendorID: vendorID, productID: productID, data: data)
+        let path = args?["path"] as? String ?? "asd"
+        printData(vendorID: vendorID, productID: productID, data: data,path: path)
     case "isConnected":
         let args = call.arguments as? [String: Any]
         let vendorID = args?["vendorId"] as? String
         let productID = args?["productId"] as? String
-   
-        // let device = findUSBDevice(vendorID: UInt16(vendorID!)!, productID: UInt16(productID!)!)
-        // if device != nil {
-        //     result(true)
-        // }else{
-        //     result(false)
-        // }
-        result(true)
+        result(connectPrinter(vendorID: vendorID!, productID: productID!))
     default:
       result(FlutterMethodNotImplemented)
     }
   }
-
-    func findUSBDevice(vendorID: UInt16, productID: UInt16) -> io_service_t? {
-        var matchingDict: NSMutableDictionary
-        matchingDict = IOServiceMatching(kIOUSBDeviceClassName)
-
-        matchingDict[kUSBVendorID] = vendorID
-        matchingDict[kUSBProductID] = productID
-
-        var iterator: io_iterator_t = 0
-        let result = IOServiceGetMatchingServices(kIOMasterPortDefault, matchingDict, &iterator)
-
-        guard result == KERN_SUCCESS else {
+    
+    public func getAllUsbDevice() -> Array<USBDevice>{
+        var devices = Array<USBDevice>()
+        let matchingDict = IOServiceMatching(kIOUSBDeviceClassName)
+        let iter = UnsafeMutablePointer<io_iterator_t>.allocate(capacity: 1)
+        let kr = IOServiceGetMatchingServices(kIOMasterPortDefault, matchingDict, iter)
+        if kr != KERN_SUCCESS {
+            print("Error getting services")
+            return devices
+        }    
+        while case let usbDevice = IOIteratorNext(iter.pointee), usbDevice != 0 {
+            devices.append(usbDevice.getInfo()!)
+        }
+        return devices
+    }
+    
+    public func connectPrinter(vendorID: String, productID: String)-> Bool{
+        let devices = getAllUsbDevice()
+        for device in devices {
+            if String(device.vendorId) == vendorID && String(device.productId) == productID {
+                return true
+            }
+        }
+        return false
+    }
+    
+    
+    func getBSDPath(for usbDevice: io_service_t) -> String? {
+        guard let bsdNameAsCFString = IORegistryEntryCreateCFProperty(usbDevice, kIOBSDNameKey as CFString, kCFAllocatorDefault, 0)?.takeRetainedValue()else {
+            print("Unable to get BSD path for USB device.")
             return nil
         }
-
-        return iterator
+        let bsdPath = bsdNameAsCFString as? String
+        print("BSD Path: \(String(describing: bsdPath))")
+        return bsdPath
     }
-
-    public func printData(vendorID: String, productID: String, data: Array<Int>) {
-        let device = findUSBDevice(vendorID: UInt16(vendorID)!, productID: UInt16(productID)!)
-        // Print Data on the printer
-        if device != nil {
-            
-            
-        } else { 
-        }
-    }
- 
-    public func isConnected(vendorID: String, productID: String) -> Bool {
-        let device = findUSBDevice(vendorID: UInt16(vendorID)!, productID: UInt16(productID)!)
-        if device != nil {
-            return true
-        }else{
-            return false
-        }
-    }
-     
-
-    public func connectPrinter(vendorID: String, productID: String) -> String {
-        var printerName: String = ""
-        let matchingDictionary = IOServiceMatching(kIOUSBDeviceClassName)
-        var iterator: io_iterator_t = 0
-        let kr = IOServiceGetMatchingServices(kIOMasterPortDefault, matchingDictionary, &iterator)
-        if kr == kIOReturnSuccess {
-            var usbDevice = IOIteratorNext(iterator)
-            while usbDevice != 0 {
-                let deviceVendorIDAsCFNumber = IORegistryEntrySearchCFProperty(usbDevice,
-                                                                   kIOServicePlane,
-                                                                   "idVendor" as CFString,
-                                                                   kCFAllocatorDefault,
-                                                                   IOOptionBits(kIORegistryIterateRecursively))
-                if deviceVendorIDAsCFNumber != nil {
-                    let deviceVendorID = deviceVendorIDAsCFNumber! as! Int
-                    if deviceVendorID == Int(vendorID) {
-                        let deviceProductIDAsCFNumber = IORegistryEntrySearchCFProperty(usbDevice,
-                                                                   kIOServicePlane,
-                                                                   "idProduct" as CFString,
-                                                                   kCFAllocatorDefault,
-                                                                   IOOptionBits(kIORegistryIterateRecursively))
-                        if deviceProductIDAsCFNumber != nil {
-                            let deviceProductID = deviceProductIDAsCFNumber! as! Int
-                            if deviceProductID == Int(productID) {
-                                printerName = getUsbDeviceName(usbDevice: usbDevice)
-                                break
-                            }
-                        }
-                    }
+    public func printData(vendorID: String, productID: String, data: Array<Int>, path: String){
+                let serialPort = ORSSerialPort(path: path)
+                serialPort?.baudRate = 9600
+                serialPort?.open()
+                var data = Data()
+                for byte in data {
+                    data.append(byte)
                 }
-                usbDevice = IOIteratorNext(iterator)
-            }
-            print("Printer Name: \(printerName) is Connected")
-        } else {
-            // Handle error
-        }
-        IOObjectRelease(iterator)
-        return printerName
+                serialPort?.send(data as Data)
+                return
+ 
     }
-    
-    public func getAllUsbDevice() -> [[String: Any]]{
-        var usbDevices: [[String: Any]] = []
-//        // Get All USB Devices connected to the system
-//        let matchingDictionary = IOServiceMatching(kIOUSBDeviceClassName)
-//        var iterator: io_iterator_t = 0
-//        let kr = IOServiceGetMatchingServices(kIOMasterPortDefault, matchingDictionary, &iterator)
-//        if kr == kIOReturnSuccess {
-//            var usbDevice = IOIteratorNext(iterator)
-//            while usbDevice != 0 {
-//                usbDevices.append(getDeviceDetails(usbDevice: usbDevice))
-//                usbDevice = IOIteratorNext(iterator)
-//            }
-//        } else {
-//            // Handle error
-//        }
-//        IOObjectRelease(iterator)
-        let devices = USBDevice.all;
-        
-        return usbDevices
-    }
-
-    public func getUsbDeviceName(usbDevice: io_service_t) -> String {
-        var deviceName: String = ""
-        let deviceNameAsCFString = IORegistryEntrySearchCFProperty(usbDevice,
-                                                                   kIOServicePlane,
-                                                                   "USB Product Name" as CFString,
-                                                                   kCFAllocatorDefault,
-                                                                   IOOptionBits(kIORegistryIterateRecursively))
-        if deviceNameAsCFString != nil {
-            deviceName = deviceNameAsCFString! as! String
-        }
-        return deviceName
-    }
-
-    // Get Device Details
-    public func getDeviceDetails(usbDevice: io_service_t) -> [String: Any] {
-        var deviceDetails: [String: Any] = [:]
-        var deviceName: String = ""
-        var deviceVendorID: Int = 0
-        var deviceProductID: Int = 0 
-
-        let deviceNameAsCFString = IORegistryEntrySearchCFProperty(usbDevice,
-                                                                   kIOServicePlane,
-                                                                     "USB Product Name" as CFString,
-                                                                   kCFAllocatorDefault,
-                                                                   IOOptionBits(kIORegistryIterateRecursively))
-        if deviceNameAsCFString != nil {
-            deviceName = deviceNameAsCFString! as! String
-        }
-        
-        let deviceVendorIDAsCFNumber = IORegistryEntrySearchCFProperty(usbDevice,
-                                                                   kIOServicePlane,
-                                                                   "idVendor" as CFString,
-                                                                   kCFAllocatorDefault,
-                                                                   IOOptionBits(kIORegistryIterateRecursively))
-        if deviceVendorIDAsCFNumber != nil {
-            deviceVendorID = deviceVendorIDAsCFNumber! as! Int
-        }
-
-        let deviceProductIDAsCFNumber = IORegistryEntrySearchCFProperty(usbDevice,
-                                                                   kIOServicePlane,
-                                                                   "idProduct" as CFString,
-                                                                   kCFAllocatorDefault,
-                                                                   IOOptionBits(kIORegistryIterateRecursively))
-        if deviceProductIDAsCFNumber != nil {
-            deviceProductID = deviceProductIDAsCFNumber! as! Int
-        }
-
-        deviceDetails["name"] = deviceName
-        deviceDetails["vendorID"] = deviceVendorID
-        deviceDetails["productID"] = deviceProductID
-        return deviceDetails
-    }
-    
 }